CustomArmour -> ArmourCustom
