+++ src/Addons/34thPRC_ArmourCustom/$PREFIX$
@@ -1,0 +1,1 @@
+34thPRC_ArmourCustom
