+++ src/Addons/34thPRC_ArmourCustom/config.cpp
@@ -1,0 +1,20 @@
+class CfgPatches
+{
+	class 34thPRC_ArmourCustom
+	{
+		author="34th PRC Modding Team";
+		units[]={};
+		requiredAddons[]=
+		{
+			"34thPRC_ArmourStandard" //34th PRC Aux Mod
+		};
+	};
+};
+class CfgVehicles
+{
+	#include "data\config_vehicles.hpp"
+};
+class CfgWeapons
+{
+	#include "data\config_weapons.hpp"
+};
