+++ src/Addons/34thPRC_ArmourCustom/data/config_weapons.hpp
