--- src/Addons/34thPRC_CustomArmour/$PREFIX$
@@ -1,1 +1,0 @@
-34thPRC_CustomArmour
