--- src/Addons/34thPRC_CustomArmour/config.cpp
@@ -1,24 +1,0 @@
-class CfgPatches
-{
-	class 34thPRC_CustomArmour
-	{
-		author="34th PRC Modding Team";
-		units[]={};
-		requiredAddons[]=
-		{
-			"34thPRC_ArmourStandard" //34th PRC Aux Mod
-		};
-	};
-};
-class CfgVehicles
-{
-	#include "data\config_vehicles.hpp"
-};
-class CfgWeapons
-{
-	#include "data\config_weapons.hpp"
-};
-class CfgGlasses
-{
-	#include "data\config_glasses.hpp"
-};
