--- src/Addons/34thPRC_CustomArmour/data/config_vehicles.hpp
@@ -1,1 +1,0 @@
-#include "odst\config_vehicles.hpp"
