--- src/Addons/34thPRC_CustomArmour/data/odst/config_vehicles.hpp
@@ -1,1 +1,0 @@
-//Put Custom ODST Rucksacks here
